clean-ups, add comments
