+++ desktop/app/app.go
@@ -99,7 +99,12 @@ type App struct {
 }
 
 // NewApp creates a new desktop app that initializes the app and acts as a moderator between all desktop components.
-func NewApp(flags flashlight.Flags, configDir string) *App {
+func NewApp() *App {
+	// initialize app config and flags based on environment variables
+	flags := initializeAppConfig()
+	configDir := flags.ConfigDir
+	initLogging(configDir)
+
 	ss := settings.LoadSettings(configDir)
 	statsTracker := NewStatsTracker()
 	app := &App{
