+++ desktop/app/config.go
@@ -4,14 +4,21 @@ import (
 	"context"
 	"encoding/json"
 	"os"
+	"path/filepath"
 	"strconv"
 	"sync"
 
+	"github.com/getlantern/appdir"
+	"github.com/getlantern/flashlight/v7"
 	fcommon "github.com/getlantern/flashlight/v7/common"
 	"github.com/getlantern/flashlight/v7/config"
 	"github.com/getlantern/lantern-client/desktop/ws"
 	"github.com/getlantern/lantern-client/internalsdk/common"
 	"github.com/getlantern/lantern-client/internalsdk/protos"
+)
+
+const (
+	defaultConfigDirPerm = 0750
 )
 
 type configService struct {
@@ -106,4 +113,53 @@ func (app *App) sendConfigOptions() {
 			OnBoardingStatus:     false,
 		},
 	})
+}
+
+// initializeAppConfig initializes application configuration and flags based on environment variables
+func initializeAppConfig() flashlight.Flags {
+	flags := flashlight.ParseFlags()
+	if flags.Pprof {
+		go startPprof("localhost:6060")
+	}
+	parseBoolEnv := func(key string, defaultValue bool) bool {
+		val := os.Getenv(key)
+		parsedValue, err := strconv.ParseBool(val)
+		if err != nil {
+			return defaultValue
+		}
+		return parsedValue
+	}
+
+	// helper to resolve CONFIG_DIR to an absolute path
+	resolveConfigDir := func(dir string) string {
+		if filepath.IsAbs(dir) {
+			return dir
+		}
+		absPath, err := filepath.Abs(dir)
+		if err != nil {
+			return dir
+		}
+		return absPath
+	}
+
+	// Parse environment-based flags
+	stickyConfig := parseBoolEnv("STICKY_CONFIG", false)
+	readableConfig := parseBoolEnv("READABLE_CONFIG", true)
+	configDir := os.Getenv("CONFIG_DIR")
+	if configDir == "" {
+		configDir = appdir.General(common.DefaultAppName)
+		log.Debugf("CONFIG_DIR not set. Using default: %s", configDir)
+	} else {
+		configDir = resolveConfigDir(configDir)
+	}
+	if err := createDirIfNotExists(configDir, defaultConfigDirPerm); err != nil {
+		log.Errorf("Unable to create config directory %s: %v", configDir, err)
+	}
+	flags.StickyConfig = stickyConfig
+	flags.ReadableConfig = readableConfig
+	flags.ConfigDir = configDir
+
+	log.Debugf("Config options: directory %v sticky %v readable %v", configDir,
+		stickyConfig, readableConfig)
+	return flags
 }
